+++ StartUp.tcl
@@ -73,7 +73,10 @@ namespace eval ::osvvm {
   variable ToolExecutable [info nameofexecutable]
   variable ToolExecutableName [file rootname [file tail $ToolExecutable]]
 
-  if {[info exists aldec]} {
+  if {[info exists ::env(OSVVM_TOOL)]} {
+    variable ScriptBaseName $::env(OSVVM_TOOL)
+    
+  } elseif {[info exists aldec]} {
     variable ToolFamily [lindex [split [vsim -version]] 2]
     if {$ToolFamily eq "Riviera-PRO"} { 
       variable ScriptBaseName   "RivieraPro"
@@ -90,6 +93,18 @@ namespace eval ::osvvm {
     
   } elseif {$ToolExecutableName eq "vish" || $ToolExecutableName eq "vsimk"} {
     variable ScriptBaseName "Siemens"
+
+    if {![catch {vsimId} msg]} {
+      variable ToolVersion [vsimId]
+    } else {
+      set VersionString [exec vsim -version &2>1]
+      regexp {(vsim\s+)(\d+\.\d+\S*)} $VersionString s1 s2 s3
+      variable ToolVersion $s3
+    }
+
+    if {[expr [string compare $ToolVersion "2025.2"] >= 0]} {
+      variable ScriptBaseName "Questa"
+    }
     if {[info exists ::env(OSVVM_TOOL)]} {
       if {$::env(OSVVM_TOOL) eq "Visualizer"} {
         variable ScriptBaseName $::env(OSVVM_TOOL)
@@ -104,9 +119,10 @@ namespace eval ::osvvm {
 
   } else {
     variable ScriptBaseName "GHDL"
-    if {[info exists ::env(OSVVM_TOOL)]} {
-      variable ScriptBaseName $::env(OSVVM_TOOL)
-    }
+    # now done at beginning
+    # if {[info exists ::env(OSVVM_TOOL)]} {
+    #   variable ScriptBaseName $::env(OSVVM_TOOL)
+    # }
   }
 }